# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Evolvus Swagger API",
    "description": "Evolvus Spec for Swagger",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "api/user": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Get an user from the database.",
        "description": "",
        "operationId": "getUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "User API"
        ],
        "summary": "Add a new user to the database",
        "description": "",
        "operationId": "addUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be added to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "api/application": {
      "get": {
        "tags": [
          "Application API"
        ],
        "summary": "Get an application from the database.",
        "description": "",
        "operationId": "getApplication",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "Application API"
        ],
        "summary": "Add a new application to the database",
        "description": "",
        "operationId": "addApplication",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Application object that needs to be added to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Application"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/application/{applicationCode}": {
      "put": {
        "tags": [
          "Application API"
        ],
        "summary": "Update an application to the database",
        "description": "",
        "operationId": "updateApplication",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Application object that needs to be updated to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "required": [
        "tenantId",
        "userName",
        "userId",
        "contact",
        "role",
        "entityId"
      ],
      "tags": [
        {
          "tenantId": "string",
          "userName": "string",
          "userId": "string",
          "contact": "object",
          "role": "object",
          "entityId": "string",
          "designation": "string",
          "masterTimeZone": "string",
          "masterCurrency": "string",
          "individualTransactionLimit": "number",
          "dailyLimit": "number"
        }
      ],
      "properties": {
        "tenantId": {
          "type": "string",
          "example": "T001"
        },
        "userName": {
          "type": "string",
          "example": "kavyak"
        },
        "userId": {
          "type": "string",
          "example": "KAVYAK"
        },
        "contact": {
          "type": "object",
          "example": {
            "phoneNumber": "9090909090",
            "mobileNumber": "8898998987",
            "emailId": "kavya.k@evolvussolutions.com",
            "country": "India",
            "fax": "",
            "state": "Karnataka",
            "city": "Hassan"
          }
        },
        "role": {
          "type": "object",
          "example": {
            "roleName": "Developer"
          }
        },
        "entityId": {
          "type": "string",
          "example": "H001B001"
        },
        "designation": {
          "type": "string",
          "example": "Engineer"
        },
        "masterTimeZone": {
          "type": "string",
          "example": "IST"
        },
        "masterCurrency": {
          "type": "string",
          "example": "INR"
        },
        "individualTransactionLimit": {
          "type": "number",
          "example": 100000
        },
        "dailyLimApplicationit": {
          "type": "number",
          "example": 1000
        }
      }
    },
    "Application": {
      "type": "object",
      "required": [
        "tenantId",
        "applicationName",
        "applicationCode",
        "entityId"
      ],
      "tags": [
        {
          "tenantId": "string",
          "entityId": "string",
          "description": "string",
          "applicationName": "string",
          "applicationCode": "string",
          "logo": "string",
          "favicon": "string"
        }
      ],
      "properties": {
        "tenantId": {
          "type": "string",
          "example": "T001"
        },
        "applicationName": {
          "type": "string",
          "example": "SANDSTORM"
        },
        "applicationCode": {
          "type": "string",
          "example": "SANDSTORM"
        },
        "entityId": {
          "type": "string",
          "example": "H001B001"
        },
        "description": {
          "type": "string",
          "example": "This is application Sandstorm"
        }
      }
    },
...
  }
}